requirements.txt
streamlit==1.42.0
openai==1.3.7
pandas==2.2.0
python-dotenv==1.0.1
pandasai==1.5.5
matplotlib==3.8.2
bs4==0.0.2
pyyaml
polars
